AI Agents/Practice_language_bot/requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
python-dotenv
openai
